Fix style issues
- Fix file system
- Replace almost all node creating functions with analogs from AST.h
- Fix brackets placing
