+++ libs/compiler/builder.c
@@ -72,8 +72,8 @@ static item_t usual_arithmetic_conversions(node *const LHS, node *const RHS)
 	return TYPE_INTEGER;
 }
 
-static node fold_unary_expression(builder *const bldr, const item_t type, const category_t ctg
-	, node *const expr, const unary_t op, const location loc)
+static node fold_unary_expression(builder *const bldr, const item_t type, const category_t ctg, node *const expr,
+								  const unary_t op, const location loc)
 {
 	if (expression_get_class(expr) != EXPR_LITERAL)
 	{
@@ -135,8 +135,8 @@ static node fold_unary_expression(builder *const bldr, const item_t type, const 
 	}
 }
 
-static node fold_binary_expression(builder *const bldr, const item_t type
-	, node *const LHS, node *const RHS, const binary_t op, const location loc)
+static node fold_binary_expression(builder *const bldr, const item_t type, node *const LHS, node *const RHS,
+								   const binary_t op, const location loc)
 {
 	if (expression_get_class(LHS) != EXPR_LITERAL || expression_get_class(RHS) != EXPR_LITERAL)
 	{
@@ -238,8 +238,8 @@ static node fold_binary_expression(builder *const bldr, const item_t type
 }
 
 
-static size_t evaluate_args(builder *const bldr, const node *const format_str
-	, item_t *const format_types, item_t *const actual_types, char32_t *const placeholders)
+static size_t evaluate_args(builder *const bldr, const node *const format_str, item_t *const format_types,
+							item_t *const actual_types, char32_t *const placeholders)
 {
 	const size_t str_index = expression_literal_get_string(format_str);
 	const char *const string = string_get(bldr->sx, str_index);
@@ -282,8 +282,8 @@ static size_t evaluate_args(builder *const bldr, const node *const format_str
 					break;
 
 				case 'p':
-				case U'у': 
-					format_types[args] = actual_types[args]; 
+				case U'у':
+					format_types[args] = actual_types[args];
 					args++;
 					break;
 
@@ -304,7 +304,7 @@ static size_t evaluate_args(builder *const bldr, const node *const format_str
 	return args;
 }
 
-static node build_printf_expression(builder *const bldr, node *const callee, node_vector *const args, const location r_loc)
+node build_printf_expression(builder *const bldr, node *const callee, node_vector *const args, const location r_loc)
 {
 	const size_t argc = node_vector_size(args);
 	if (args == NULL || argc == 0)
@@ -331,12 +331,12 @@ static node build_printf_expression(builder *const bldr, node *const callee, nod
 	{
 		node argument = node_vector_get(args, i);
 		item_t arg_type = expression_get_type(&argument);
-		actual_types[i-1] = arg_type;
+		actual_types[i - 1] = arg_type;
 	}
 
 	char32_t placeholders[MAX_PRINTF_ARGS];
 	item_t format_types[MAX_PRINTF_ARGS];
-	const size_t expected_args = evaluate_args(bldr, &fst, format_types, actual_types, placeholders); 
+	const size_t expected_args = evaluate_args(bldr, &fst, format_types, actual_types, placeholders);
 
 	if (expected_args != argc - 1)
 	{
@@ -358,14 +358,15 @@ static node build_printf_expression(builder *const bldr, node *const callee, nod
 
 	const location loc = { node_get_location(callee).begin, r_loc.end };
 	return expression_call(TYPE_INTEGER, callee, args, loc);
-} 
+}
 
-static node build_print_expression(builder *const bldr, node *const callee, node_vector *const args, const location r_loc)
+static node build_print_expression(builder *const bldr, node *const callee, node_vector *const args,
+								   const location r_loc)
 {
-	const location loc = { node_get_location(callee).begin, r_loc.end };    
-	
-	node_vector stmts = node_vector_create();    
-		
+	const location loc = { node_get_location(callee).begin, r_loc.end };
+
+	node_vector stmts = node_vector_create();
+
 	node_vector args_to_print = node_vector_create();
 
 	size_t argc;
@@ -373,48 +374,52 @@ static node build_print_expression(builder *const bldr, node *const callee, node
 	{
 		semantic_error(bldr, r_loc, expected_expression);
 		return node_broken();
-	} 
+	}
 
-	bool complicated_type_in_args = 0;  
+	bool complicated_type_in_args = 0;
 
 	size_t last_scalar_argument_index = 0;
 	size_t first_scalar_argument_index = 0;
 	size_t last_argument_index = 0;
 
 	vector str = vector_create(1);
-	
-	for (size_t i = 0; i < argc; i++)
-	{  
-		node argument = node_vector_get(args, i); 
 
-		location curr_loc = node_get_location(&argument); 
-		
+	for (size_t i = 0; i < argc; i++)
+	{
+		node argument = node_vector_get(args, i);
+
+		location curr_loc = node_get_location(&argument);
+
 		node last_argument = *callee;
 		if (i > 0)
 		{
 			last_argument_index = i - 1;
 			last_argument = node_vector_get(args, last_argument_index);
-		} 
-		location last_argument_loc = node_get_location(&last_argument);  
-		
-		item_t argument_type = expression_get_type(&argument);  
-		if ((type_is_aggregate(bldr->sx, argument_type) && !type_is_string(bldr->sx, argument_type)) || 
+		}
+		location last_argument_loc = node_get_location(&last_argument);
+
+		item_t argument_type = expression_get_type(&argument);
+		if ((type_is_aggregate(bldr->sx, argument_type) && !type_is_string(bldr->sx, argument_type)) ||
 			(type_is_pointer(bldr->sx, argument_type)))
-		{  
-			complicated_type_in_args = 1;  
-			
+		{
+			complicated_type_in_args = 1;
+
 			if (type_is_array(bldr->sx, argument_type))
 			{
 				if (!vector_add_str(&str, "{"))
+				{
 					return node_broken();
+				}
 			}
 			else if (type_is_structure(bldr->sx, argument_type))
 			{
-				if (!vector_add_str(&str, "{ struct")) 
+				if (!vector_add_str(&str, "{ struct"))
+				{
 					return node_broken();
-			}  
+				}
+			}
 
-			// корректно определяем location	
+			// корректно определяем location
 			node first_scalar_argument = node_vector_get(args, first_scalar_argument_index);
 			location fs_arg_loc = node_get_location(&first_scalar_argument);
 			node last_scalar_argument = node_vector_get(args, last_scalar_argument_index);
@@ -423,22 +428,22 @@ static node build_print_expression(builder *const bldr, node *const callee, node
 			// разворачиваемся в printf для имеющейся на данный момент строки
 			if (vector_size(&str))
 			{
-				node printf_node = create_printf_node_by_vector(bldr, &str, &args_to_print, fs_arg_loc, ls_arg_loc); 
-				// запоминаем узел 
-				node_vector_add(&stmts, &printf_node); 
+				node printf_node = create_printf_node_by_vector(bldr, &str, &args_to_print, fs_arg_loc, ls_arg_loc);
+				// запоминаем узел
+				node_vector_add(&stmts, &printf_node);
 			}
 
 			vector_clear(&str);
 
 			// дальнейшая строка и аргументы не будут иметь к только что построенному узлу никакого отношения
 			str = vector_create(1);
-			// избавляемся от предыдущих запомненных аргументов 
+			// избавляемся от предыдущих запомненных аргументов
 			node_vector_clear(&args_to_print);
-			args_to_print = node_vector_create(); 
-			first_scalar_argument_index = 0; 
-			last_scalar_argument_index = 0; 
-			
-			node complicated_type_node = create_complicated_type_str(bldr, &argument, last_argument_loc, curr_loc, 1); 
+			args_to_print = node_vector_create();
+			first_scalar_argument_index = 0;
+			last_scalar_argument_index = 0;
+
+			node complicated_type_node = create_complicated_type_str(bldr, &argument, last_argument_loc, curr_loc, 1);
 			if (!node_is_correct(&complicated_type_node))
 			{
 				vector_clear(&str);
@@ -446,59 +451,75 @@ static node build_print_expression(builder *const bldr, node *const callee, node
 			}
 			node_vector_add(&stmts, &complicated_type_node);
 
-			if (type_is_array(bldr->sx, argument_type)) 
+			if (type_is_array(bldr->sx, argument_type))
 			{
 				if (!vector_add_str(&str, "} "))
+				{
 					return node_broken();
+				}
 			}
 			else if (type_is_structure(bldr->sx, argument_type))
 			{
 				if (!vector_add_str(&str, "\n}\n"))
+				{
 					return node_broken();
+				}
 			}
-			else if (i != argc-1)
+			else if (i != argc - 1)
 			{
 				if (!vector_add_str(&str, " "))
+				{
 					return node_broken();
-			} 
+				}
+			}
 		}
 		else
-		{ 
+		{
 			if (last_scalar_argument_index == 0)
+			{
 				first_scalar_argument_index = i;
+			}
 
-			last_scalar_argument_index = i; 
-			
+			last_scalar_argument_index = i;
+
 			if (!type_is_null_pointer(argument_type))
+			{
 				node_vector_add(&args_to_print, &argument);
+			}
 			else
+			{
 				node_remove(&argument);
+			}
 
 			const item_t argument_type_class = type_get_class(bldr->sx, argument_type);
-			const char *tmp = create_simple_type_str(argument_type_class);   
+			const char *tmp = create_simple_type_str(argument_type_class);
 			if (!tmp)
 			{
 				vector_clear(&str);
-				return node_broken(); 
+				return node_broken();
 			}
 			if (!vector_add_str(&str, tmp))
+			{
 				return node_broken();
-			if (i != argc-1)
+			}
+			if (i != argc - 1)
 			{
 				if (!vector_add_str(&str, " "))
+				{
 					return node_broken();
+				}
 			}
-		}  
-	}  
+		}
+	}
 	node last_argument = node_vector_get(args, first_scalar_argument_index);
-	location last_argument_loc = node_get_location(&last_argument); 
+	location last_argument_loc = node_get_location(&last_argument);
 
 	node_remove(callee);
 
-	// разворачиваемся в printf для имеющейся на данный момент строки 
+	// разворачиваемся в printf для имеющейся на данный момент строки
 	if (vector_size(&str))
 	{
-		node printf_node = create_printf_node_by_vector(bldr, &str, &args_to_print, last_argument_loc, r_loc);   
+		node printf_node = create_printf_node_by_vector(bldr, &str, &args_to_print, last_argument_loc, r_loc);
 
 		if (!complicated_type_in_args)
 		{
@@ -507,18 +528,19 @@ static node build_print_expression(builder *const bldr, node *const callee, node
 			return printf_node;
 		}
 
-		node_vector_add(&stmts, &printf_node); 
+		node_vector_add(&stmts, &printf_node);
 	}
 	vector_clear(&str);
-		 
-	return expression_inline(TYPE_VOID, &stmts, loc); 
-} 
 
-static node build_printid_expression(builder *const bldr, node *const callee, node_vector *const args, const location r_loc)
-{ 
-	const location loc = { node_get_location(callee).begin, r_loc.end };   
-	
-	node_vector stmts = node_vector_create();    
+	return expression_inline(TYPE_VOID, &stmts, loc);
+}
+
+static node build_printid_expression(builder *const bldr, node *const callee, node_vector *const args,
+									 const location r_loc)
+{
+	const location loc = { node_get_location(callee).begin, r_loc.end };
+
+	node_vector stmts = node_vector_create();
 
 	node_vector args_to_print = node_vector_create();
 
@@ -527,16 +549,16 @@ static node build_printid_expression(builder *const bldr, node *const callee, no
 	{
 		semantic_error(bldr, r_loc, expected_identifier_in_printid);
 		return node_broken();
-	} 
+	}
 
-	bool complicated_type_in_args = 0;    
+	bool complicated_type_in_args = 0;
 
 	size_t last_scalar_argument_index = 0;
 	size_t first_scalar_argument_index = 0;
 	size_t last_argument_index = 0;
-	
+
 	vector str = vector_create(1);
- 
+
 	for (size_t i = 0; i < argc; i++)
 	{
 		node argument = node_vector_get(args, i);
@@ -544,48 +566,58 @@ static node build_printid_expression(builder *const bldr, node *const callee, no
 		{
 			semantic_error(bldr, node_get_location(&argument), expected_identifier_in_printid);
 			return node_broken();
-		}  
+		}
 		// аргумент -- идентификатор, а узел -- нет => идентификатор не объявлен, но
 		// об этом уже объявлено => можем выходить из цикла
-		if (node_get_type(&argument) != OP_IDENTIFIER) 
+		if (node_get_type(&argument) != OP_IDENTIFIER)
+		{
 			break;
+		}
 
-		location curr_loc = node_get_location(&argument); 
- 
+		location curr_loc = node_get_location(&argument);
+
 		node last_argument = *callee;
 		if (i > 0)
 		{
 			last_argument_index = i - 1;
 			last_argument = node_vector_get(args, last_argument_index);
-		} 
-		location last_argument_loc = node_get_location(&last_argument); 
+		}
+		location last_argument_loc = node_get_location(&last_argument);
 
 		// добавляем строку "имя_аргумента = "
 		const size_t argument_identifier_index = expression_identifier_get_id(&argument);
-		const char *tmp_argument_name = ident_get_spelling(bldr->sx, argument_identifier_index); 
+		const char *tmp_argument_name = ident_get_spelling(bldr->sx, argument_identifier_index);
 		if (!vector_add_str(&str, tmp_argument_name))
+		{
 			return node_broken();
+		}
 		if (!vector_add_str(&str, " = "))
+		{
 			return node_broken();
-		
-		item_t argument_type = expression_get_type(&argument);  
-		if ((type_is_aggregate(bldr->sx, argument_type) && !type_is_string(bldr->sx, argument_type)) || 
+		}
+
+		item_t argument_type = expression_get_type(&argument);
+		if ((type_is_aggregate(bldr->sx, argument_type) && !type_is_string(bldr->sx, argument_type)) ||
 			(type_is_pointer(bldr->sx, argument_type)))
-		{  
-			complicated_type_in_args = 1;  
-			
+		{
+			complicated_type_in_args = 1;
+
 			if (type_is_array(bldr->sx, argument_type))
 			{
 				if (!vector_add_str(&str, "{"))
+				{
 					return node_broken();
+				}
 			}
 			else if (type_is_structure(bldr->sx, argument_type))
 			{
 				if (!vector_add_str(&str, "{ struct"))
+				{
 					return node_broken();
+				}
 			}
 
-			// корректно определяем location	
+			// корректно определяем location
 			node first_scalar_argument = node_vector_get(args, first_scalar_argument_index);
 			location fs_arg_loc = node_get_location(&first_scalar_argument);
 			node last_scalar_argument = node_vector_get(args, last_scalar_argument_index);
@@ -594,22 +626,22 @@ static node build_printid_expression(builder *const bldr, node *const callee, no
 			// разворачиваемся в printf для имеющейся на данный момент строки
 			if (vector_size(&str))
 			{
-				node printf_node = create_printf_node_by_vector(bldr, &str, &args_to_print, fs_arg_loc, ls_arg_loc); 
-				// запоминаем узел 
+				node printf_node = create_printf_node_by_vector(bldr, &str, &args_to_print, fs_arg_loc, ls_arg_loc);
+				// запоминаем узел
 				node_vector_add(&stmts, &printf_node);
 			}
 
-			vector_clear(&str); 
+			vector_clear(&str);
 
-			// дальнейшая строка и аргументы не будут иметь к только что построенному узлу никакого отношения 
+			// дальнейшая строка и аргументы не будут иметь к только что построенному узлу никакого отношения
 			str = vector_create(1);
-			// избавляемся от предыдущих запомненных аргументов 
+			// избавляемся от предыдущих запомненных аргументов
 			node_vector_clear(&args_to_print);
-			args_to_print = node_vector_create(); 
-			first_scalar_argument_index = 0; 
-			last_scalar_argument_index = 0; 
-			
-			node complicated_type_node = create_complicated_type_str(bldr, &argument, last_argument_loc, curr_loc, 1); 
+			args_to_print = node_vector_create();
+			first_scalar_argument_index = 0;
+			last_scalar_argument_index = 0;
+
+			node complicated_type_node = create_complicated_type_str(bldr, &argument, last_argument_loc, curr_loc, 1);
 			if (!node_is_correct(&complicated_type_node))
 			{
 				vector_clear(&str);
@@ -617,56 +649,68 @@ static node build_printid_expression(builder *const bldr, node *const callee, no
 			}
 			node_vector_add(&stmts, &complicated_type_node);
 
-			if (type_is_array(bldr->sx, argument_type)) 
+			if (type_is_array(bldr->sx, argument_type))
 			{
 				if (!vector_add_str(&str, "} "))
+				{
 					return node_broken();
+				}
 			}
 			else if (type_is_structure(bldr->sx, argument_type))
 			{
 				if (!vector_add_str(&str, "\n}\n"))
+				{
 					return node_broken();
+				}
 			}
-			else if (i != argc-1)
+			else if (i != argc - 1)
 			{
 				if (!vector_add_str(&str, " "))
+				{
 					return node_broken();
+				}
 			}
 		}
 		else
-		{  
+		{
 			if (last_scalar_argument_index == 0)
+			{
 				first_scalar_argument_index = i;
+			}
 
-			last_scalar_argument_index = i; 
-			
+			last_scalar_argument_index = i;
+
 			node_vector_add(&args_to_print, &argument);
 
 			const item_t argument_type_class = type_get_class(bldr->sx, argument_type);
-			const char *tmp = create_simple_type_str(argument_type_class);   
+			const char *tmp = create_simple_type_str(argument_type_class);
 			if (!tmp)
 			{
 				vector_clear(&str);
-				return node_broken(); 
+				return node_broken();
 			}
 			if (!vector_add_str(&str, tmp))
+			{
 				return node_broken();
-			if (i != argc-1)
+			}
+			if (i != argc - 1)
 			{
 				if (!vector_add_str(&str, " "))
+				{
 					return node_broken();
+				}
 			}
-		} 	 
-	}  
+		}
+	}
 	node last_argument = node_vector_get(args, first_scalar_argument_index);
-	location last_argument_loc = node_get_location(&last_argument);  
- 
+	location last_argument_loc = node_get_location(&last_argument);
+
 	node_remove(callee);
- 
-	// разворачиваемся в printf для имеющейся на данный момент строки 
+
+	// разворачиваемся в printf для имеющейся на данный момент строки
 	if (vector_size(&str))
 	{
-		node printf_node = create_printf_node_by_vector(bldr, &str, &args_to_print, last_argument_loc, r_loc);    
+		node printf_node = create_printf_node_by_vector(bldr, &str, &args_to_print, last_argument_loc, r_loc);
 
 		if (!complicated_type_in_args)
 		{
@@ -675,14 +719,15 @@ static node build_printid_expression(builder *const bldr, node *const callee, no
 			return printf_node;
 		}
 
-		node_vector_add(&stmts, &printf_node); 
+		node_vector_add(&stmts, &printf_node);
 	}
 	vector_clear(&str);
-		 
-	return expression_inline(TYPE_VOID, &stmts, loc); 
-} 
 
-static node build_getid_expression(builder *const bldr, node *const callee, node_vector *const args, const location r_loc)
+	return expression_inline(TYPE_VOID, &stmts, loc);
+}
+
+static node build_getid_expression(builder *const bldr, node *const callee, node_vector *const args,
+								   const location r_loc)
 {
 	const size_t argc = node_vector_size(args);
 	if (args == NULL || argc == 0)
@@ -870,8 +915,8 @@ node build_string_literal_expression(builder *const bldr, const size_t index, co
 	return expression_string_literal(&bldr->context, type_string(bldr->sx), index, loc);
 }
 
-node build_subscript_expression(builder *const bldr, node *const base, node *const index
-	, const location l_loc, const location r_loc)
+node build_subscript_expression(builder *const bldr, node *const base, node *const index, const location l_loc,
+								const location r_loc)
 {
 	if (!node_is_correct(base) || !node_is_correct(index))
 	{
@@ -898,8 +943,8 @@ node build_subscript_expression(builder *const bldr, node *const base, node *con
 	return expression_subscript(element_type, base, index, loc);
 }
 
-node build_call_expression(builder *const bldr, node *const callee
-	, node_vector *const args, const location l_loc, const location r_loc)
+node build_call_expression(builder *const bldr, node *const callee, node_vector *const args, const location l_loc,
+						   const location r_loc)
 {
 	if (!node_is_correct(callee))
 	{
@@ -956,8 +1001,8 @@ node build_call_expression(builder *const bldr, node *const callee
 	return expression_call(return_type, callee, args, loc);
 }
 
-node build_member_expression(builder *const bldr, node *const base, const size_t name
-	, const bool is_arrow, const location op_loc, const location id_loc)
+node build_member_expression(builder *const bldr, node *const base, const size_t name, const bool is_arrow,
+							 const location op_loc, const location id_loc)
 {
 	if (!node_is_correct(base))
 	{
@@ -1051,8 +1096,8 @@ node build_unary_expression(builder *const bldr, node *const operand, const unar
 	const item_t operand_type = expression_get_type(operand);
 
 	const location loc = op_kind == UN_POSTINC || op_kind == UN_POSTDEC
-		? (location){ node_get_location(operand).begin, op_loc.end }
-		: (location){ op_loc.begin, node_get_location(operand).end };
+							 ? (location){ node_get_location(operand).begin, op_loc.end }
+							 : (location){ op_loc.begin, node_get_location(operand).end };
 
 	switch (op_kind)
 	{
@@ -1151,8 +1196,8 @@ node build_unary_expression(builder *const bldr, node *const operand, const unar
 	}
 }
 
-node build_binary_expression(builder *const bldr, node *const LHS, node *const RHS
-	, const binary_t op_kind, const location op_loc)
+node build_binary_expression(builder *const bldr, node *const LHS, node *const RHS, const binary_t op_kind,
+							 const location op_loc)
 {
 	if (!node_is_correct(LHS) || !node_is_correct(RHS))
 	{
@@ -1240,9 +1285,8 @@ node build_binary_expression(builder *const bldr, node *const LHS, node *const R
 				return fold_binary_expression(bldr, TYPE_BOOLEAN, LHS, RHS, op_kind, loc);
 			}
 
-			if ((type_is_pointer(bldr->sx, left_type) && type_is_null_pointer(right_type))
-				|| (type_is_null_pointer(left_type) && type_is_pointer(bldr->sx, right_type))
-				|| left_type == right_type)
+			if ((type_is_pointer(bldr->sx, left_type) && type_is_null_pointer(right_type)) ||
+				(type_is_null_pointer(left_type) && type_is_pointer(bldr->sx, right_type)) || left_type == right_type)
 			{
 				return fold_binary_expression(bldr, TYPE_BOOLEAN, LHS, RHS, op_kind, loc);
 			}
@@ -1346,8 +1390,7 @@ node build_ternary_expression(builder *const bldr, node *const cond, node *const
 		return expression_ternary(LHS_type, cond, LHS, RHS, loc);
 	}
 
-	if ((type_is_null_pointer(LHS_type) && type_is_pointer(bldr->sx, RHS_type))
-		|| (LHS_type == RHS_type))
+	if ((type_is_null_pointer(LHS_type) && type_is_pointer(bldr->sx, RHS_type)) || (LHS_type == RHS_type))
 	{
 		return expression_ternary(RHS_type, cond, LHS, RHS, loc);
 	}
@@ -1396,8 +1439,8 @@ node build_empty_bound_expression(builder *const bldr, const location loc)
 }
 
 
-node build_member_declaration(builder *const bldr, const item_t type, const size_t name, const bool was_star
-	, node_vector *const bounds, const location loc)
+node build_member_declaration(builder *const bldr, const item_t type, const size_t name, const bool was_star,
+							  node_vector *const bounds, const location loc)
 {
 	if (type_is_void(type))
 	{
@@ -1461,8 +1504,8 @@ node build_struct_declaration(builder *const bldr, node *const declaration, node
 	return declaration_struct_set_location(declaration, loc);
 }
 
-node build_declarator(builder *const bldr, const item_t type, const size_t name
-   , const bool was_star, node_vector *const bounds, node *const initializer, const location ident_loc)
+node build_declarator(builder *const bldr, const item_t type, const size_t name, const bool was_star,
+					  node_vector *const bounds, node *const initializer, const location ident_loc)
 {
 	if (type_is_void(type))
 	{
@@ -1592,8 +1635,8 @@ node build_null_statement(builder *const bldr, const location semi_loc)
 	return statement_null(&bldr->context, semi_loc);
 }
 
-node build_if_statement(builder *const bldr, node *const cond
-	, node *const then_stmt, node *const else_stmt, const location if_loc)
+node build_if_statement(builder *const bldr, node *const cond, node *const then_stmt, node *const else_stmt,
+						const location if_loc)
 {
 	if (!node_is_correct(cond) || !node_is_correct(then_stmt) || (else_stmt && !node_is_correct(else_stmt)))
 	{
@@ -1661,11 +1704,11 @@ node build_do_statement(builder *const bldr, node *const body, node *const cond,
 	return statement_do(body, cond, loc);
 }
 
-node build_for_statement(builder *const bldr, node *const init
-	, node *const cond, node *const incr, node *const body, const location for_loc)
+node build_for_statement(builder *const bldr, node *const init, node *const cond, node *const incr, node *const body,
+						 const location for_loc)
 {
-	if ((init && !node_is_correct(init)) || (cond && !node_is_correct(cond))
-		|| (incr && !node_is_correct(incr)) || !node_is_correct(body))
+	if ((init && !node_is_correct(init)) || (cond && !node_is_correct(cond)) || (incr && !node_is_correct(incr)) ||
+		!node_is_correct(body))
 	{
 		return node_broken();
 	}
